+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Time.hs
@@ -12,7 +12,10 @@ module Plutarch.LedgerApi.V1.Time (
 
 import Plutarch.Builtin (PDataNewtype (PDataNewtype))
 import Plutarch.DataRepr (DerivePConstantViaData (DerivePConstantViaData))
-import Plutarch.Enum (PCountable (psuccessor, psuccessorN))
+import Plutarch.Enum (
+  PCountable (psuccessor, psuccessorN),
+  PEnumerable (ppredecessor, ppredecessorN),
+ )
 import Plutarch.LedgerApi.Utils (Mret)
 import Plutarch.Lift (
   PConstantDecl,
@@ -54,6 +57,15 @@ instance PCountable PPosixTime where
   psuccessorN = phoistAcyclic $ plam $ \p t ->
     let p' = pcon . PPosixTime . pcon . PDataNewtype . pdata . pto $ p
      in p' + t
+
+-- | @since WIP
+instance PEnumerable PPosixTime where
+  {-# INLINEABLE ppredecessor #-}
+  ppredecessor = phoistAcyclic $ plam (\t -> t - 1)
+  {-# INLINEABLE ppredecessorN #-}
+  ppredecessorN = phoistAcyclic $ plam $ \p t ->
+    let p' = pcon . PPosixTime . pcon . PDataNewtype . pdata . pto $ p
+     in p' - t
 
 -- | @since 2.0.0
 instance PIntegral PPosixTime where